+++ packages/mobile/alpha/CardMedia.tsx
@@ -1,12 +1,13 @@
 import { createCardMedia } from '@cbhq/cds-common/cards/createCardMedia';
 
-import { Illustration } from '../illustrations/Illustration';
+import { Pictogram, SpotSquare } from '../illustrations';
 
 import { CardRemoteImage } from './CardRemoteImage';
 
 export type CardMediaProps = React.ComponentProps<typeof CardMedia>;
 
 export const CardMedia = createCardMedia({
+  SpotSquare,
+  Pictogram,
   CardRemoteImage,
-  Illustration,
 });