+++ packages/mobile/illustrations/Illustration.tsx
@@ -19,9 +19,7 @@ type IllustrationProps<Type extends IllustrationVariant> = IllustrationBaseProps
 };
 
 /**
- * @deprecated Do not use this component directly as this is not optimized for tree-shaking and will cause all
- * illustration assets to be included in your app. This component is used within CDS for visreg.
- *
+ * @deprecated This component will be deleted in Q42023. Please use HeroSquare, SpotSquare, SpotRectangle, or Pictogram components instead.
  */
 export const Illustration = memo(function Illustration<Type extends IllustrationVariant>({
   type,