+++ packages/mobile/illustrations/__stories__/HeroSquare.stories.tsx
@@ -1,3 +1,3 @@
 import { getIllustrationSheet } from './getIllustrationSheet';
 
-export default getIllustrationSheet('heroSquare', { dimension: '200x200', scaleMultiplier: 0.24 }); // render a 48x48 thumbnail
+export default getIllustrationSheet('heroSquare');