+++ packages/mobile/illustrations/__stories__/Pictogram.stories.tsx
@@ -1,3 +1,3 @@
 import { getIllustrationSheet } from './getIllustrationSheet';
 
-export default getIllustrationSheet('pictogram', { dimension: '48x48' });
+export default getIllustrationSheet('pictogram');