+++ packages/mobile/illustrations/__stories__/SpotRectangle.stories.tsx
@@ -1,6 +1,3 @@
 import { getIllustrationSheet } from './getIllustrationSheet';
 
-export default getIllustrationSheet('spotRectangle', {
-  dimension: '240x120',
-  scaleMultiplier: 0.3, // render a 72x36 thumbnail
-});
+export default getIllustrationSheet('spotRectangle');