+++ packages/mobile/illustrations/__stories__/SpotSquare.stories.tsx
@@ -1,3 +1,3 @@
 import { getIllustrationSheet } from './getIllustrationSheet';
 
-export default getIllustrationSheet('spotSquare', { dimension: '96x96', scaleMultiplier: 0.4 }); // render a 48x48 thumbnail
+export default getIllustrationSheet('spotSquare');