+++ packages/web/alpha/CardMedia.tsx
@@ -1,6 +1,6 @@
 import { createCardMedia } from '@cbhq/cds-common/cards/createCardMedia';
 
-import { Illustration } from '../illustrations/Illustration';
+import { Pictogram, SpotSquare } from '../illustrations';
 
 import { CardRemoteImage } from './CardRemoteImage';
 
@@ -8,5 +8,6 @@ export type CardMediaProps = React.ComponentProps<typeof CardMedia>;
 
 export const CardMedia = createCardMedia({
   CardRemoteImage,
-  Illustration,
+  SpotSquare,
+  Pictogram,
 });