+++ packages/web/illustrations/__stories__/getIllustrationSheet.tsx
@@ -1,5 +1,12 @@
 import React, { memo } from 'react';
-import type { IllustrationNamesMap, IllustrationVariant, Spectrum } from '@cbhq/cds-common';
+import type {
+  HeroSquareDimension,
+  IllustrationNamesMap,
+  IllustrationVariant,
+  PictogramDimension,
+  Spectrum,
+} from '@cbhq/cds-common';
+import { illustrationDimensions, illustrationSizes } from '@cbhq/cds-common/tokens/illustrations';
 import heroSquareVersionMap from '@cbhq/cds-illustrations/__generated__/heroSquare/data/versionMap';
 import pictogramVersionMap from '@cbhq/cds-illustrations/__generated__/pictogram/data/versionMap';
 import spotRectangleVersionMap from '@cbhq/cds-illustrations/__generated__/spotRectangle/data/versionMap';
@@ -10,10 +17,39 @@ import { VStack } from '../../layout';
 import { FeatureFlagProvider } from '../../system';
 import { ThemeProvider } from '../../system/ThemeProvider';
 import { TextLegal } from '../../typography';
-import { Illustration } from '../Illustration';
+import {
+  HeroSquare,
+  HeroSquareName,
+  Pictogram,
+  PictogramName,
+  SpotRectangle,
+  SpotRectangleName,
+  SpotSquare,
+  SpotSquareName,
+} from '../index';
 
-const ITEM_SIZE = 48;
-const ITEM_DIMENSION = `${ITEM_SIZE}x${ITEM_SIZE}` as const;
+const SIZES = {
+  pictogram: {
+    height: illustrationSizes[illustrationDimensions.pictogram[0]][1],
+    width: illustrationSizes[illustrationDimensions.pictogram[0]][0],
+    scaleMultiplier: 1,
+  },
+  spotSquare: {
+    height: illustrationSizes[illustrationDimensions.spotSquare[0]][1],
+    width: illustrationSizes[illustrationDimensions.spotSquare[0]][0],
+    scaleMultiplier: 0.4,
+  },
+  spotRectangle: {
+    height: illustrationSizes[illustrationDimensions.spotRectangle[0]][1],
+    width: illustrationSizes[illustrationDimensions.spotRectangle[0]][0],
+    scaleMultiplier: 0.3,
+  },
+  heroSquare: {
+    height: illustrationSizes[illustrationDimensions.heroSquare[0]][1],
+    width: illustrationSizes[illustrationDimensions.heroSquare[0]][0],
+    scaleMultiplier: 0.24,
+  },
+};
 
 function keys<T>(obj: { [key in keyof T]: T[key] }) {
   return Object.keys(obj) as unknown as Extract<keyof T, string>[];
@@ -46,6 +82,7 @@ export function getIllustrationSheet<Type extends IllustrationVariant>(type: Typ
   });
 
   const renderItem = ({ name, spectrum }: DataItem) => {
+    const dim = `${SIZES[type].width}x${SIZES[type].height}` as const;
     return (
       <ThemeProvider scale="xSmall" spectrum={spectrum}>
         <VStack
@@ -53,11 +90,39 @@ export function getIllustrationSheet<Type extends IllustrationVariant>(type: Typ
           background
           gap={1}
           alignItems="flex-start"
-          width={ITEM_SIZE}
-          height={ITEM_SIZE + 20}
+          width={SIZES[type].width * SIZES[type].scaleMultiplier}
+          height={SIZES[type].height * SIZES[type].scaleMultiplier + 20}
           overflow="hidden"
         >
-          <Illustration type={type} name={name} dimension={ITEM_DIMENSION} />
+          {type === 'heroSquare' && (
+            // render a 48x48 thumbnail
+            <HeroSquare
+              name={name as HeroSquareName}
+              dimension={dim as HeroSquareDimension}
+              scaleMultiplier={SIZES.heroSquare.scaleMultiplier}
+            />
+          )}
+          {type === 'spotSquare' && (
+            // render a 48x48 thumbnail
+            <SpotSquare
+              name={name as SpotSquareName}
+              scaleMultiplier={SIZES.spotSquare.scaleMultiplier}
+            />
+          )}
+          {type === 'spotRectangle' && (
+            // render a 72x36 thumbnail
+            <SpotRectangle
+              name={name as SpotRectangleName}
+              scaleMultiplier={SIZES.spotRectangle.scaleMultiplier}
+            />
+          )}
+          {type === 'pictogram' && (
+            <Pictogram
+              name={name as PictogramName}
+              dimension={dim as PictogramDimension}
+              scaleMultiplier={SIZES.pictogram.scaleMultiplier}
+            />
+          )}
           <TextLegal as="p" noWrap>
             {name}
           </TextLegal>