+++ packages/web/CHANGELOG.md
@@ -8,7 +8,15 @@ All notable changes to this project will be documented in this file!
 
 <!-- template-start -->
 
-## Unreleased
+## 5.8.0 (8/26/2023 PST)
+
+#### 🚀 Updates
+
+- Update CardMedia to use SpotSquare & Pictogram components. [[#2200](https://github.cbhq.net/frontend/cds/pull/2200)] [[DX-813](https://jira.coinbase-corp.com/browse/DX-813)]
+
+#### 📘 Misc
+
+- Updated tests to use illustration componnents directly and deleted Illustration test. [[#2200](https://github.cbhq.net/frontend/cds/pull/2200)] [[DX-813](https://jira.coinbase-corp.com/browse/DX-813)]
 
 #### 📘 Misc